Skip existing

+++ mkpkg.sh
@@ -24,9 +24,16 @@ else
 		cd $cwd/.tmp/$cmd
 
 		clib() {
+			[ -e lib/$1 ] || [ -e lib64/$1 ] && return;
 			echo -n "$1 "
 			[ -e $PREFIX/lib64/$1 ] && cp $PREFIX/lib64/$1 lib64
 			[ -e $PREFIX/lib/$1 ] && cp $PREFIX/lib/$1 lib
+			[ -e /lib/$1 ] && cp /lib/$1 lib
+			[ -e /lib64/$1 ] && cp /lib64/$1 lib64
+			[ -e $PREFIX/local/lib/$1 ] && cp $PREFIX/local/lib/$1 lib
+			[ -e $PREFIX/local/lib64/$1 ] && cp $PREFIX/local/lib64/$1 lib64
+			[ -e /system/lib/$1 ] && cp /system/lib/$1 lib
+			[ -e /system/lib64/$1 ] && cp /system/lib64/$1 lib64
 		}
 
 		for i in lib lib64 bin; do
